refactor: remove legacy campaign code
